Author_post

+++ admin/includes/view_all_posts.php
@@ -91,12 +91,12 @@ if (isset($_POST['checkboxArray'])){
         $post_comment_count = $row['post_comment_count'];
         $post_date = $row['post_date'];
 
-
         echo "<tr>";
-        ?>
-        <td><input type='checkbox' class='checkBoxes' name="checkboxArray[]" value="<?php echo $post_id; ?>"></td>;
 
-                      <?php
+        ?>
+        <td><input type='checkbox' class='checkBoxes' name="checkboxArray[]" value='<?php echo $post_id; ?>'></td>;
+
+        <?php
         echo "<td> {$post_id}</td>";
         echo "<td>{$post_author} </td>";
         echo "<td>{$post_title} </td>";
@@ -121,7 +121,7 @@ if (isset($_POST['checkboxArray'])){
         echo "<td>{$post_date}</td>";
         echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+        echo "<td><a onclick=\"javascript: return confirm('Are you sure you want to Delete');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
 
         echo "</tr>";
